pass C2 build control flag (BUILD_AARCH64_C2=true) through to hotspot make
this is needed so that the top level makefile can drive the AAARCH64
C2 compile when building the hotspot aarch64_c2 branch. since it is
orthogonal to the hotspot C1 and default branch builds it is ok for it
to go into the top level Makefile in the default branch.